# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAText API
  description: |
    API specification for WASAText messaging application.
    Built according to the PDF specification.
  version: "1.0.0"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Use the user identifier returned from doLogin
  schemas:
    User:
      type: object
      description: Representation of a user in the system
      properties:
        identifier:
          type: string
          description: Unique user identifier (UUID hex string)
          example: "abcdef012345"
          minLength: 1
          maxLength: 64
          pattern: '^[a-zA-Z0-9_-]+$'
        name:
          type: string
          description: Username chosen by the user (3-16 characters)
          minLength: 3
          maxLength: 16
          pattern: '^[a-zA-Z0-9_-]+$'
          example: "Maria"
        photo:
          type: string
          format: binary
          description: User profile photo in binary format (optional)
...
    Error:
      type: object
      description: Standard error response object
      properties:
        message:
          type: string
          description: Human-readable explanation of what went wrong
          example: "Username already taken"
          minLength: 1
          maxLength: 256
          pattern: '^[\s\S]+$'
  parameters:
    UserId:
      name: userId
      in: path
      description: Targeted user identifier
      required: true
      schema:
        type: string
        description: User identifier value
        minLength: 1
        maxLength: 64
        pattern: '^[a-zA-Z0-9_-]+$'
...
paths:
  /session:
    post:
      tags: ["login"]
      summary: Logs in the user
      description: |
        If the user does not exist, it will be created,
        and an identifier is returned.
        If the user exists, the user identifier is returned.
      operationId: doLogin
      requestBody:
        description: User nickname for identification
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Sign-in request body
              properties:
                name:
                  type: string
                  description: Nickname chosen by the user
                  example: Maria
                  pattern: '^[a-zA-Z0-9_-]+$'
                  minLength: 3
                  maxLength: 16
              required:
                - name
      responses:
        '201':
          description: User log-in action successful
          content:
            application/json:
              schema:
                type: object
                description: Successful login response
                properties:
                  identifier:
                    type: string
                    description: The unique identifier assigned to the user
                    example: "abcdef012345"
                    minLength: 1
                    maxLength: 64
                    pattern: '^[a-zA-Z0-9_-]+$'
  /users/{userId}/username:
    parameters:
      - $ref: '#/components/parameters/UserId'
    put:
      tags: ["user"]
      summary: Update the user's username
      description: |
        Users can update their name, provided the new name 
        is not already in use by someone else.
      operationId: setMyUserName
      security:
        - bearerAuth: []
      requestBody:
        description: The new nickname to set
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Username update request
              properties:
                name:
                  type: string
                  description: New nickname
                  minLength: 3
                  maxLength: 16
                  pattern: '^[a-zA-Z0-9_-]+$'
                  example: "NewName"
              required:
                - name
      responses:
        '200':
          description: Username updated successfully
          content:
            application/json:
              schema:
                type: object
                description: Empty successful response
                properties:
                  status:
                    type: string
                    description: Operation result status
                    example: "success"
                    minLength: 1
                    maxLength: 16
                    pattern: '^[a-zA-Z]+$'
        '400':
          description: Invalid username format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Username already taken
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/{userId}/photo:
    parameters:
      - $ref: '#/components/parameters/UserId'
    put:
      tags: ["user"]
      summary: Set the user's profile photo
      description: Upload or update the user's profile photo.
      operationId: setMyPhoto
      security:
        - bearerAuth: []
      requestBody:
        description: The image file data
        required: true
        content:
          image/*:
            schema:
              type: string
              format: binary
              description: Binary image data
      responses:
        '200':
          description: Photo updated successfully
          content:
            application/json:
              schema:
                type: object
                description: Empty successful response
                properties:
                  status:
                    type: string
                    description: Operation result status
                    example: "success"
                    minLength: 1
                    maxLength: 16
                    pattern: '^[a-zA-Z]+$'
        '400':
          description: Invalid photo format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users:
    get:
      tags: ["user"]
      summary: Search for users by username
      description: |
        User can search for other users via the username 
        and see all the existing WASAText usernames.
      operationId: searchUsers
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
            description: Search query string
            minLength: 1
            maxLength: 64
            pattern: '^[a-zA-Z0-9_-]*$'
          description: Username to search for (partial match)
      responses:
        '200':
          description: List of matching users
          content:
            application/json:
              schema:
                type: array
                description: Array of users matching the search
                minItems: 0
                maxItems: 100
                items:
                  $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...